Get rid of gi_composites
The gi_composites module  allowed constructing widgets from templates
but was only used for the main window and headerbar. Having composite
templates for complicated widgets makes sense but in this case it was
just being used for trivial purposes, something that did not need a
lot of boilerplate code using the gtk builder way. When the feature
actually lands in PyGI, it will hopefully support nested templates
and might help separate UI abstractions from the implementation for
good.