Use GtkListBox with GioListStore as model for notes
It is not possible to have title, summary and probably one or two
buttons in in the same row using CellRenderers. For this kind of
design, a GtkListBox is needed, with a Gio.ListStore as model
representing notes from a group.

Leaving notesview.ListView and store.TreeStore in place, to be used
later for tree view of note groups (aka notebooks).