Fix preview not working for initial load
Switching to a note for the first time while in preview mode
would not work because it reads content to render before it is
loaded into the buffer. After making some changes to the file API
it is possible to call `load_file` after async read has loaded the
actual contents of the file into the buffer. The active buffer's
`buffer-changed` signal needs to be blocked meanwhile.