Introduce word goals and remove markup and overwrite mode labels
The markup and overwrite mode indicators are disctracting. Markup
can be checked and set (for later sessions) from a separate section
meant for document preferences. Overwrite mode is indicated by the
editor itself (fat cursor bar). Word count is still useful and
something that the writer would like to keep an eye on.

The UI for setting word goals has been implemented as well, but
is quite useless since it needs to be preserved over sessions.