Make title and subtitle updates to DB, non-blocking for UI
Consecutive updates to the DB from the same thread as GTK main loop
freeze the UI and make it unresponsive until all operations have
been executed. A straightforward fix is doing the updates in a
separate thread. This can, however, cause very long queues which
need a longer time for connections waiting for transactional locks
to be lifted. A timeout of 60 seconds seems reasonable, but a
proper solution would be to support indefinite amount of queueing.