Rename application to 'Draft'
The name 'Noto' was originally meant to denote the purpose of note
taking. But a proper note taking app would generally provide more
than just markup based text editing. Changing the purpose of this
project to before its 0.1 release indicates some lack of vision on
the author's part, but for the better (or worse), this application
will now be crafted around the goal of providing writers a tool for
creating digital documents. For this purpose - Draft - seems like
a more fitting name.
reason, the new name - Draft - seems more fitting.