Ensure group creation only when named
Last commit created and then deleted group when a valid name was
not given. It makes more sense to create group only when it has
been named in the first place. Also this way, safeguards against
the scenario where a group has been created and not yet named but
the process quits - leaving an undesired group in DB when it should
have been deleted.