Implement basic drag and drop for group and texts
Moving texts and groups to other groups has been a goal for Draft
for a long time. This commit enables this, but in a somewhat
limited way. Possible extensions of this feature should be multi
selection drags (to move multiple groups and text at once), scroll
while dragging and finally drag and drop to rearrange order of
groups or texts within their respective views.