Implement proper shortcut based multi selection in text list
Two shortcuts `Ctrl + Primary` and `Shift + Primary` are generally
used for selecting multiple items from a collection. The `Ctrl`
modifier helps in enlargening the selection and the `Shift` key
allows selecting all the rows from the last selected row to the
row being pointed to. This multi-selection feature only helps the
user to transfer multiple items from one group to another. Editing
is forbidden when more than one items in the list are selected.