Allow creating models having texts based on condition type
DraftTextListModel is able to view a list of texts but this need
not be limited to 'texts from a specific group'. The same view can
be used for providing lists that have been trashed or were modified
recently. This commit makes five different text models posiible:
texts belonging to a group, texts that have been labelled with some
tag, texts recently modified, texts that have been trashed and
finally a list of all non-trashed texts in the database.