Hide link-address associated with links in the buffer
Links and images can have somewhat long and distracting URLs which
can be hidden from the user and allowed to edit interactively. For
now these URLS are just hidden (but not using invisibility text
tag - owing to some issues in GtkTextView and invisible characters)
but will be followed up with a UI for editing links.