Use inherent multi-selection for DraftBaseist
The `activate-on-single-click` property when set to fasle does
seem to allow for proper multi-selectom within ListBox. So the
custom multi-selection mechanisms can be discarded. It should be
noted that this would not work well if single click activation is
enabled for some purpose, but that is unlikely, considering the
interaction design for Draft.