Major architectural change - make list model reactive
Up until now, any changes made by the editor to a sheet (text) was
conveyed to the DraftTextList view, which in turn requested its
model to update the database. This increased complexity and put
extraneous business logic within a view widget (DraftTextList). A
reactive row item on the other hand can emit signals when its data
is changed thus, allowing a handler to push changes to DB and move
files if necessary, to reflect that change. Making changes to the
DB can now be triggered by simply assigning a value to any of the
relevant properties of a TextRowData instance.